+++ lib/pages/journal_page.dart
@@ -149,12 +149,10 @@ class _JournalPageState extends State<JournalPage> {
       queryStyle: const TextStyle(
         fontFamily: 'Lato',
         fontSize: 24,
-        fontWeight: FontWeight.w300,
       ),
       hintStyle: const TextStyle(
         fontFamily: 'Lato',
         fontSize: 24,
-        fontWeight: FontWeight.w300,
       ),
       physics: const BouncingScrollPhysics(),
       borderRadius: BorderRadius.circular(8.0),