+++ lib/pages/settings/measurables/measurables_page.dart
@@ -76,12 +76,10 @@ class _MeasurablesPageState extends State<MeasurablesPage> {
       queryStyle: const TextStyle(
         fontFamily: 'Lato',
         fontSize: 20,
-        fontWeight: FontWeight.w300,
       ),
       hintStyle: const TextStyle(
         fontFamily: 'Lato',
         fontSize: 20,
-        fontWeight: FontWeight.w300,
       ),
       physics: const BouncingScrollPhysics(),
       borderRadius: BorderRadius.circular(8.0),