+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added:
-- Use Romanian localization
+- Romanian localization
 
 ## [0.8.17] - 2022-05-22
 ### Fixed: